r10965/sofa : push 59f019e..e5f60bd
* commit e5f60bd716c59cedbcf8efeb499a3a47a14b99ff
Resurrect doxygen documentation

This commit adds the possibility generate doxygen documentation on a
per-project basis, in a way that is integrated with CMake. As it turns
out, cutting the documentation in chunks yields significant improvement
in generation time.

The different documentations are linked together, yet they are separate.
This means there is also a separate "main page" for each project, which
could be a nice place to write documentation, but which is blank for the
moment.

To build the documentation, try:
$ make doc
and open doc/SOFA/index.html

Feedback welcome.

As a side note, to work around the file organisation in modules/, the
cmake script tries to extract the source files of each module through
ungodly use of unix utilities, so Windows users are out of luck for now.

   6 files changed, 3222 insertions(+), 0 deletions(-)




Former-commit-id: 4d9c38becf3bb1b71501a69a68d554dfbd69ab97