Add etc/sofa.ini configuration file
This file is generated by CMake. It is written in a basic ini style, and
contains only the paths to the share/ and examples/ directories.

For now, it always points to the source tree, but the idea is that once
we can install Sofa, it will point to the installed directories.


Former-commit-id: 696567871d901853d63baf16aa850612e7035470