cmake: Add a relative rpath to installed binaries
The rpath is a value stored in the header of a binary (for example an
ELF binary), which is a hint for the system loader of where to find the
needed shared objects.

This commit uses CMAKE_INSTALL_RPATH to add a '../lib' relative rpath in
the installed versions of the shared libraries and executables.

Having a relative rpath allows us to have fully relocatable binaries.
I don't know if '../lib' this is enough, and I don't know if it works on
OS X.

For reference, more info about RPATH handling with CMake:

  http://www.cmake.org/Wiki/CMake_RPATH_handling


Former-commit-id: 487374bf41bb9f586e9421d9f4a271094baeebca