[CMakeList.txt] Add two new config options to allow us to handle software evolution
Software evolution is hard an requires care and strategy.
To ease this we have splitted the set of sofa files in three sets:
- Minimal (The component of SofaNG)
- Full
- Deprecated

The two added options works in the following way:
When SOFA_BUILD_FULLSETCOMPONENTS is ON the components in full are added minimal
When SOFA_BUILD_DEPRECATEDCOMPONENTS is ON the components in the deprecated set are added to full set.

This organization is orthogonal to Plugins which are more specific to "context"
specific component (like haptic, implicitfield or whatever).