[SofaGeneralLoader] Make GridMeshCreator work again (#2473)
3 scenes were displaying an error following the PR #2465 .

After some investigations, it appeared that those 3 scenes were not working for a while (and the PR merely showed that statement), and the culprit was that GridMeshCreator was not calling its doLoad() method.
doLoad() in MeshLoader components is triggered when d_filename is set.
But GridMeshCreator  is procedural and does not any filename at all. Therefore doLoad() was never called.
This PR forcefully set the data d_filename to dirty to force-trigger the callback.

Moreover, insertTriangle and insertQuad were using beginEdit()/endEdit() which were triggering also a doLoad() (resulting in a lot of useless edges/triangles)
This PR uses WriteOnlyAccessor to avoid this problem.